requirements.txt
cffi==2.0.0
colorama==0.4.6
contourpy==1.3.3
crayons==0.4.0
cryptography==46.0.3
cycler==0.12.1
fernet==1.0.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
piexif==1.1.3
pillow==11.3.0
pyaes==1.6.1
pybind11==3.0.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
stegano==2.0.0
